tests/handlers/ServerHandler.test.ts
import { ServerHandler } from '../../src/handlers/ServerHandler';
import { SystemInfo } from '../../src/types';

// Mock the fs module since ServerHandler uses it
jest.mock('fs', () => ({
  existsSync: jest.fn().mockReturnValue(true),
}));

// Create a mock class based on the abstract ServerHandler to test its concrete methods
class MockServerHandler extends ServerHandler {
    executeCommand(command: string, timeout?: number): Promise<{ stdout: string; stderr: string }> {
        // Check if the command is to generate a large response
        if (command === 'generate-large-response') {
          // Generate a large string to simulate a large response
          const largeResponse = 'a'.repeat(5000); // Adjust the size as needed for testing
          return Promise.resolve({ stdout: largeResponse, stderr: '' });
        }
        // Mock implementation for other commands
        return Promise.resolve({ stdout: 'Command executed', stderr: '' });
      }
  setCurrentDirectory(directory: string): boolean {
    // Mock implementation
    return true;
  }
  getCurrentDirectory(): Promise<string> {
    // Mock implementation
    return Promise.resolve('/mock/directory');
  }
  listFiles(directory: string, limit?: number, offset?: number, orderBy?: string): Promise<string[]> {
    // Mock implementation
    return Promise.resolve(['file1.txt', 'file2.txt', 'file3.txt']);
  }
  createFile(directory: string, filename: string, content: string, backup: boolean): Promise<boolean> {
    // Mock implementation
    return Promise.resolve(true);
  }
  updateFile(filePath: string, pattern: string, replacement: string, backup: boolean): Promise<boolean> {
    // Mock implementation
    return Promise.resolve(true);
  }
  amendFile(filePath: string, content: string): Promise<boolean> {
    // Mock implementation
    return Promise.resolve(true);
  }
  getSystemInfo(): Promise<SystemInfo> {
    // Complete mock implementation according to the SystemInfo type
    const mockSystemInfo: SystemInfo = {
      freeMemory: 1024,
      totalMemory: 2048,
      homeFolder: '/mock/home', // Add mock value
      type: 'MockType', // Add mock value
      release: 'MockRelease', // Add mock value
      pythonVersion: '3.8.5', // Mock Python version
      cpuArchitecture: 'x64', // Mock CPU architecture
      uptime: 123456, // Mock system uptime in seconds
      currentFolder: '/mock/current', // Mock current folder
      platform: 'mockPlatform' // Add mock value
    };
  
    return Promise.resolve(mockSystemInfo);
  }
}

describe('ServerHandler', () => {
  let mockServerHandler: MockServerHandler;

  beforeEach(() => {
    mockServerHandler = new MockServerHandler();
  });

  describe('Pagination', () => {
    it('should handle large responses from executeCommand', async () => {
      const { stdout } = await mockServerHandler.executeCommand('generate-large-response');
      const responseId = mockServerHandler['storeResponse'](stdout);

      expect(mockServerHandler['responseStorage']).toHaveProperty(responseId);
      expect(mockServerHandler['responseStorage'][responseId].length).toBeGreaterThan(1);
    });    
    
    it('should store responses correctly', () => {
      const longResponse = 'a'.repeat(5000); // Create a response string longer than maxResponseSize
      const responseId = mockServerHandler['storeResponse'](longResponse);

      expect(mockServerHandler['responseStorage']).toHaveProperty(responseId);
      expect(mockServerHandler['responseStorage'][responseId].length).toBeGreaterThan(1);
    });

    it('should retrieve the correct page of data', () => {
      const responseString = 'a'.repeat(3000); // Create a response string to span multiple pages
      const responseId = mockServerHandler['storeResponse'](responseString);
      const pageData = mockServerHandler.getResponsePage(responseId, 1); // Retrieve the second page

      expect(pageData).not.toBeNull();
      expect(pageData?.data).toBe(responseString.substring(1024, 2048));
    });


    it('should return null when trying to retrieve a page beyond the pagination limit', () => {
      const responseString = 'a'.repeat(5000); // Ensure this is larger than your maxResponseSize
      const responseId = mockServerHandler['storeResponse'](responseString);
      
      // Calculate the number of pages
      const maxResponseSize = 1024; // This should match the maxResponseSize in your ServerHandler
      const numPages = Math.ceil(responseString.length / maxResponseSize);

      // Attempt to retrieve a page beyond the last page
      const beyondLastPageData = mockServerHandler.getResponsePage(responseId, numPages);

      expect(beyondLastPageData).toBeNull();
    });

    it('should return null for an invalid page number', () => {
      const responseString = 'a'.repeat(1024);
      const responseId = mockServerHandler['storeResponse'](responseString);
      const invalidPageData = mockServerHandler.getResponsePage(responseId, 2); // There should only be one page

      expect(invalidPageData).toBeNull();
    });
  });

  // Add more tests for other methods if needed
});

export {};
